refactor(open-mpm): split registry/mod.rs into store + tests submodules (#366)
